Updated openstack/openstack Project: openstack/requirements  9c4978ff5cc9d6f560aab790072ad8486978fc1d
Block oslo.config 3.18.0

There is a issue with tuples that was causing
ocatavia (and others to have issues with this new
release).

Since this is summit week and new releases aren't
being recommended (since they may destabalize things)
we are just going to block 3.18.0 until we can further
release a patched version (a fix has merged).

Change-Id: Ib902d57a8d5c49271b931fe8f18ec6f90159fe97
Related-Bug: #1635717