Updated openstack/openstack Project: openstack-infra/project-config  cca2f3252cd2dd0dfc173c62ca7c5762ac97e3b2
Remove Mitaka stable compat job, adjust filtering for Newton

We recently introduced a new Newton stable compat job [1] for projects
such as the oslo libraries.

As things are now, it is currently voting in stable/mitaka branch
reviews and it is probably not expected for Newton jobs to be passing
on Mitaka backports. Example of such failing backport here [2].

This change is analogous to the adjustments we did when Mitaka
released [3].

[1]: https://review.openstack.org/#/c/375111/
[2]: https://review.openstack.org/#/c/389190/
[3]: https://review.openstack.org/#/c/295500/

Change-Id: I04112e13ca69998c069c9e30f55e1f069f902253