Updated openstack/openstack Project: openstack/nova  385b04c8cc9852b11125ee7a7eca3294b5720d41
Catch error and log warning when not able to update mtimes

When we launch an instance, nova updates the mtime for the _base image
to let the image cache manager know the image is actively used.  This
can lead to unexpected I/O errors when launching a large amount of
instances at once coming from the same _base image.

This commit puts the execute call in a try, except block to catch
possible errors. It also logs a warning with the path and error message.
With this, at least once the update will succeed.

Closes-Bug: #1621818

Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>

Change-Id: I2fd1700aa4563a906eb574cbbe16caa63abae0d6
Signed-off-by: Joris S'heeren <joris.sheeren@combellgroup.com>