Updated openstack/openstack Project: openstack/oslo.context  e52b6f790f17edc321e9a1a3c41f945ed86a71f4
Fix X-User and X-Tenant deprecated headers in from_environ

The X-User and X-Tenant deprecated headers from auth_token middleware
are actually name values, not ID values. In most cases this doesn't
matter because these are deprecated and only used when no newer headers
are available but we should still fix it.

Closes-Bug: #1614776
Change-Id: I16e80671e853387004434e766dd6ceebe0a6b2a6