Updated openstack/openstack Project: openstack-infra/storyboard  d6e24ea719d97c3146df5d99f1a595b40c0e10eb
Fix modification of private stories

We introduced a weird bug where editing private stories would cause
a 500 error. This commit fixes that by not checking permissions when
trying to update the updated_at attribute of the story, which will
only ever be done once we've verified that the current user does have
permission to see the story.

Change-Id: I1a774a2d9653199e5f4c811181df365199b6c109