Updated openstack/openstack Project: openstack/nova  503856e940c81a1a1b2f3286e8cdca53ceb026c0
Fix database poison warnings, part 19

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    nova.tests.unit.virt.ironic.test_driver.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I4181775eb0e89e8a21e6af8dddb3c9ffd09a6b83
Related-Bug: #1568414