Updated openstack/openstack Project: openstack/nova  b46f074caee90967df4761115251f6050f51d032
Fix database poison warnings, part 21

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    nova.tests.unit.compute.test_compute_api.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I7bdf8422b050039df02c40a069964262dfe4a9eb
Related-Bug: #1568414