Updated openstack/openstack Project: openstack/nova-specs  504c10b2b6aaa7a0673e2450ace672b39efadca9
Websockify security proxy framework

This proposes a generic framework for supporting TLS
encryption between the console proxy servers and the
compute nodes for SPICE, VNC and serial ports.

NB, the spec content is largely rewritten to better
reflect what was actually implemented. The key change
to previous cycles is the removal of the config
parameter that let you set the classname for the
security proxy, thus indirectly enabling tls. Instead
you the config explicitly says whether you want tls or
not and nova chooses the impl to match.

Previously approved: Juno, Kilo, Mitaka, Newton
Blueprint websocket-proxy-to-host-security
Change-Id: I44b472d0907606885ff75ffce27dc3f46f6565f4