Updated openstack/openstack Project: openstack/nova  ca2690e47082a38abd0c1488258417f01f6dba7d
Fix database poison warnings, part 22

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    nova.tests.unit.api.openstack.compute.test_quotas.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I1f6d5e161abea3a66fe09159b84e100c123561bd
Related-Bug: #1568414