Updated openstack/openstack Project: openstack/python-swiftclient  481f74caf1ffc649128bb8db84ac9104304420a9
Low-level API: Don't log just before raising an exception

The only logging we should do is when we've encountered a problem *and
we've dealt with it ourselves*. When we're raising an exception, it
should be up to the caller to decide whether to log anything about it.

Anything else is just rude.

Change-Id: I1c96b76d90a78b7a10ffe63e4a7440c8f579147c
Closes-Bug:  1213179
Related-Bug: 1202229