Updated openstack/openstack Project: openstack-infra/storyboard  553bf7aa1112ad581099f384cf5532e7f1cf034b
Allow permissions to be set for teams on stories

This commit allows people to add teams to the ACL for viewing
and editing a private story. This will make managing who can
see stories much easier, since it allows grouping of users such
that only one change needs to be made if someone is added to
or removed from the group.

Previously the list of users needed to be sent every time the
story was updated, this commit fixes that such that if no list
of users is sent, no change is made to the users who can see
the story.

Story: 2000568
Task: 2985
Change-Id: Idb7f881c4d772d373b583ecc2f59231be8f7f01b