Updated openstack/openstack Project: openstack/designate  e8cdfa68ee6fb1af72dd120711178fa2223f4b46
Fix unit tests for oslo.config 3.18.0

The unit tests were improperly overriding the server list with a tuple
rather than a list. In newer versions of oslo.config they are no longer
converting the values which leads to unit test failures. This change
updates the unit tests to use a list for the server override.

Change-Id: I429417e26eb535560821cc8e31dbecc9f1d26e6d
Closes-Bug: #1635717