Updated openstack/openstack Project: openstack/neutron-fwaas  27f3bbc89a403300515883f79b6e9c94633b583c
Remove import of neutron.extensions

As I was evaluating the list of places where neutron is imported into
neutron-fwaas, this instance jumped out at me as an example of something
that was probably pulled in as part of a template once upon a time, and
is no longer necessary.  All tests succeed without it.

Change-Id: I752d801481295d8e0e36de31e93f519917289fe5