Updated openstack/openstack Project: openstack/barbican  4c19e28729d090656a69131254160e179f0ae5b4
Use http_proxy_to_wsgi middleware

This sets up the HTTPProxyToWSGI middleware in front of Barbican.
The purpose of thise middleware is to set up the request URL
correctly in case there is a proxy (For instance, a loadbalancer
such as HAProxy) in front of Barbican.

The HTTPProxyToWSGI is off by default and needs to be enabled
via a configuration value.

Change-Id: Iad0151ca41684fa2d8eb60c343028e13c3719e66
Closes-bug: #1590608