Updated openstack/openstack Project: openstack/python-ceilometerclient  c1141ed5be72ee169a7640b5d4957bfe02ad37ce
move old oslo-incubator code out of openstack/common

As part of the first community-wide goal, teams were asked
to remove the openstack/common package of their projects
if one existed. This was a byproduct of the old oslo-incubator
form of syncing common functionality.

The package, apiclient, was moved to a top level location
and a new i18n module was created. There is no oslo.apiclient
library or equivalent, the recommended solution is to move it
in tree and maintain it there.

Change-Id: Ia788313e0926dc872a87b090ef0a350898bfb079