Updated openstack/openstack Project: openstack/cinder  2351f05dcc799fe99b78b1aa91792c02f4e6697b
VMAX driver - Create initiator group in single call

The VMAX driver created the initiator group first
and then added in the initiators one-by-one. This
potentially had a performance impact because there could
be a significant number of initiators. This fix creates
the initiator group, containing all the initiators, in a
single call.

Change-Id: Ib8a2ad067442d6eef06de6888e8787af8e5bd1fc
Closes-Bug: #1555728