Updated openstack/openstack Project: openstack/cinder  51411b629907d60ef928871823728bbc23d72752
VMAX driver - No support for creating CG from CG source

The VMAX driver only supports creating a consistency group
from a consistency group snapshot. The fix throws an
exception stating that we don't support creating a CG from
a CG and fixes the inconsistent state the CG falls into as
a result of it, so the user is able to delete it.

Change-Id: I8dd61f6a281f9e8ecfafc4bf2a12efa51ae2956d
Closes-Bug: #1612344