Updated openstack/openstack Project: openstack/openstack-doc-tools  67691c3cfcbbfd151009b914be42860b2519548d
Don't include openstack/common in flake8 exclude list

There is no this directory openstack/common which was used to keep codes from
oslo-incubator, we have retired oslo-incubator. Removing openstack/commonfrom
all OpenStack code in favor of the Oslo libraries is a project wide goal
for the Ocata release. So don't use this directory any more. We should drop it
for improving searching efficiency.

Change-Id: If2ca9333319fd78a53c73ab313411ed2f17b07f2