Updated openstack/openstack Project: openstack-infra/project-config  7243ae5b940253adacdccf1e6c8b2af2f3c15b94
Make zuul_swift_upload.py more accurate for mime type detection

When a text file is quite generic like a .css or .js file
libmagic will detect the file's content as text/plain. This
patch adds a check of the mime database in order to guess the
mime type given a file extension. This is only done
if text/plain is detected by libmagic.

Change-Id: Ic73332533fc4d6c1a89ccbdb59d44f3697c56dc0