Updated openstack/openstack Project: openstack-infra/shade  43ea349586f1df12dd631a765a843e224481a0ce
Document and be more explicit in normalization

Put extra keys in both the root resource and in a properties dict.

Ensure data types are correct. Make sure int, float and bool values
are returned as int and bool.

Change disabled in flavor to is_disabled for consistency with other
bools we've added. There has been no release with the addition of disabled,
so changing it now is still safe.

Add locations and direct_url to images. They're optional in glance, but
that's evil.

Let image schema attribute fall through to extra properties.

Add zone to current_location.

Add readable mappings for power_state, task_state, vm_state, launched_at
and terminated_at for Servers. Also add a non-camel-cased host_id.

This is a big patch, but it's mostly just reorganizing and adding docs.
Looking at the changes to the tests and seeing that the only change is
adding zone and properties into a couple of fixtures is a good place to
start.

Change-Id: If5674c049c8dd85ca0b3483b7c2dc82b9e139bd6