Updated openstack/openstack Project: openstack/nova  04122258908e09883414e1f09e919c41252bcea4
Send events to all relevant hosts if migrating

Previously, external events were sent to the instance object's host
field. This patch fixes the external event dispatching to check for
migration. If an instance is being migrated, the source and
destination compute are added to the set of hosts to which the event
is sent.

Change-Id: If00736ab36df4a5a3be4f02b0a550e4bcae77b1b
Closes-bug: 1535918
Closes-bug: 1624052