Updated openstack/openstack Project: openstack/tripleo-heat-templates  0f41197e6aeded5d5d2a0c3f9f0cd1a0d709d087
Re-add NFS backend for Glance

We lost ability to store Glance images in NFS mounts as we moved to NG
HA architecture. This patch re-adds that ability, but the parameter
interface changes because the semantics change as well. (Pacemaker
allowed for different mounts than just NFS so the parameters were more
generic, although we only ever tested and documented NFS usage.)

Change-Id: Ic5197e09846bbf75d780dcc74da1717dcf8301d0
Related-Bug: #1635606