Updated openstack/openstack Project: openstack-infra/zuul  38ce39fe5808810a5b3ea6fb4ce0d951757977f1
Use separate library directories for pre and post

The custom command module used in order to collect job output was
also being used by the pre and post playbooks.  This meant that
instead of going to the ansible log file, the rsync output would
end up in /tmp/console.html on the zuul launcher.

To correct this, create separate library directories for use by
the pre and post playbooks which will contain all of the modules
except the custom command module.  Write separate ansible.cfg files
for them, and instruct ansible-playbook to use those config files.

Change-Id: I5eb6bcc48bcaa6b056af1af7da93f29408f9db41