Updated openstack/openstack Project: openstack/governance  e03e7122c013a7f773bdfc2d35fa4d893d1e5f04
Overdue update of mission statement in charter

Back in February, the TC proposed a change in the mission statement:
http://governance.openstack.org/resolutions/20160217-mission-amendment.html

That change was approved by the Board on March 29:
https://wiki.openstack.org/wiki/Governance/Foundation/29Mar2016BoardMeeting

As a result, an update of the charter was proposed and merged.
Unfortunately, it used wording from an earlier mission statement
change proposal, and nobody spotted the issue in review:
https://review.openstack.org/#/c/292610/

This change applies the latest (and approved) wording.

Change-Id: Ice1a05f4eaf3c442de823011f20f0214d4872fda