Updated openstack/openstack Project: openstack/governance  5fd98db9e217c144786644f5669ddc0ece79079f
Tighten up langauge in "Contribution Is Our Currency"

Discussion of I876f95ccbbf71f4f512275af110690888449e093 indicated
that while there was general agreement with the spirit of the
content, there were some issues with redundancy and lack of
active voice. This change tries to address some of those issues
by paring the message down to its distillation.

As I started making changes I realized there was a lot of extra
fluff. We can be direct here with what we really mean.

Change-Id: If06a763e18af9fc13e8a67721382904473d8d34b