Updated openstack/openstack Project: openstack/nova  40f9b0ad16d3a5fae184bbe6d4a49cf792967089
pci: in free_device(), compare by device id and not reference

The resource tracker might have a reference to a PCI device and fails
to free it because some of its attributes doesn't match any devices in the
PciDevTracker (for example the status of the device changed).

For this reason only match on device id when freeing a PCI device.

Co-Authored-By: Ludovic Beliveau <ludovic.beliveau@windriver.com>

Change-Id: Ifc9529d9184c57923f106d8ac1f647d5e70d2896