Updated openstack/openstack Project: openstack-infra/project-config  4f43d3e6821a9f5b2e9fd98756958e358f571ac6
Switch nova to run neutron grenade multinode jobs

Nova is working on slowly unwinding it's dependence on
nova-network in the CI system. This change makes nova
run the grenade multinode job with neutron rather than
nova-network. Nova still has grenade coverage for nova-network
for nova with the gate-grenade-dsvm-ubuntu-xenial job
that is a full upgrade.

Change-Id: Idf5e165b8df9bdb88b4fdafe76cc548e483ff891