Updated openstack/openstack Project: openstack-infra/shade  7f7cf7ef618e0c2161ec1edbe694724b6d22850b
list_servers(): thread safety: never return bogus data.

Shade has a caching mechanism which is good. It's okay to return stale
data, that is, to return an "old view" of the resources. It's not okay
to return an "impossible" set of data, a bogus view of the data.

If two threads call list_severs() at the same time, the first will
acquire a lock, and start the query to Nova. The second
will try to acquire, lock is held, so it will return the cached server
list, which initially is '[]'. But there could be some servers, we
just don't know at that point.

In that case, we should block to acquire the lock, instead of return
an invalid view of the list of servers.

How to repro (Python3)
1) Create a couple of Nova servers
2) Run
import os_client_config, shade, concurrent.futures

cloud = os_client_config.make_shade(cloud='devstack')
with concurrent.futures.ThreadPoolExecutor(4) as exec:
    for i in range(3):
        exec.submit(lambda: print(len(cloud.list_servers())))

Change-Id: I32783d291c5d3d13b2fe86cd876931e25f8c4879