Updated openstack/openstack Project: openstack/nova  ad695a2691abe2aaadbb29ee970de4ba8a74824d
libvirt: Pass Host instead of Driver to volume drivers

We were initialising libvirt volume drivers by passing the
LibvirtDriver object. However, this is only ever used to fetch the
Host. We update the interface to pass the Host instead. This is
cleaner, but also better represents the intent of the interface. They
should not be able to access arbitrary attributes of the LibvirtDriver
object.

Change-Id: I87ceeee1ec46dc22754321574b16e6c47b27a848