Updated openstack/openstack Project: openstack/ironic  d41cf18a4c47dd9eccec2882fc4e960990439d50
API: lookup() ignore malformed MAC addresses

This patch is adding code to ignore malformed MAC addresses which are
called as part of the lookup() mechanism. Prior to this patch, if a MAC
address wasn't in the expected format (six octets) the deployment of the
node would fail because the ramdisk wouldn't be able to lookup which
node it was in the Ironic database. One way to trigger this problem was
to deploy a node with an Infiniband Card which the MAC address (or GID)
contains 20 octets, that would result in a deployment failure even when
that NIC wasn't used by Ironic at all (not enrolled as a port).

The ListOfMacAdresses type was also deleted as part of this patch
because it wasn't used anywhere else.

Change-Id: I614fe63236985438d2f354d17a15d17649e72912
Closes-Bug: #1633585