Updated openstack/openstack Project: openstack/python-keystoneclient  a2ab99983e88f9d9dd34a121c800ef8ab3b98a14
Support domain-specific configuration management

Provide support for the domain-specific configuration storage available
via the REST API.

Domain configs are JSON blobs and we have fine grained control on them
via the Identity API. This fine grained control is not defined yet in the
client, though - for now, we can manage everything like Python dictionaries
and use operations like "update" whenever we want to delete a specific group
or option. This approach is similar to what is done in the federation mapping
API to handle mapping rules.

Functional tests are also included, this is useful to check if the new
feature works in an integration environment.

Co-Auhtored-By: Henry Nash <henryn@linux.vnet.ibm.com>
Co-Authored-By: Rodrigo Duarte <rduartes@redhat.com>

Closes-Bug: 1433306

Partially Implements: blueprint domain-config-ext

Change-Id: Ie6795b8633fed38c58b79250c11c9a045b7f95a4