Updated openstack/openstack Project: openstack-infra/devstack-gate  695ff9d3979d3f74ff7e7d061c0a14a73d017b3f
Make reproduce.sh header more accurate

We run on more than ubuntu trusty now, try to reflect that information
and tell people where to determine exactly what distro to run on to
properly reproduce a job. Also update the package list to include
libssl-dev as it is now needed for cryptography things.

Change-Id: I635b133b1775a6a931f520404a5fa8a23514e60f