Updated openstack/openstack Project: openstack/requirements  6e26a322eec1605133377e9b64eb762071db1e20
Block kombu 4.0.0

oslo.messaging is not compatible with it, for two reasons: because we
cap amqp<2.0.0 there; and because the code of the library needs some
adjustments.

Until we clear that up, we will block the new  kombu release. We don't
cap because it's against the direction we want to take long term
(adopting to new upstream releases).

Change-Id: I34e0ca49966d679ff60d4bcfbf7157c6289a9aef
Related-Bug: #1638263