Updated openstack/openstack Project: openstack/nova  012f822992030c98c57333c4975ff0c7e9c184e5
Cleanup before removal of conductor local apis

Some old cruft from the when conductor proxying was being used to
make db calls are left in various places in unit tests. None of this
is of use anymore and needs to be removed before the conductor local
apis and related 'use_local' flag can be removed.

Change-Id: Id5fe30f4124396ac59ef3d008972db4dd930855d