Updated openstack/openstack Project: openstack/swift  a852b1ecd747714dea0f770cb88ce3a6a0c30bf8
Make NullLogger more logger like

eventlet.wsgi.server is adaptive to dealing with the named log
argument being either a file like or a log like [1], but message
handling is a more direct pass through if it's a log like.

Since all we want to do is make eventlet's logging it a noop - it's
easy to provide either interface!

1. https://github.com/eventlet/eventlet/blob/4d2cdc/eventlet/wsgi.py#L246

Change-Id: I2d792176c96931eafb3f140e6653ba8b31eda429