Updated openstack/openstack Project: openstack-infra/project-config  4fca5d0a75ad093d78e88ad811e433ce1855fe82
Restrict gate-tempest-dsvm-full-ubuntu-xenial to newton

The gate-tempest-dsvm-full-ubuntu-xenial job uses nova-network
which is deprecated since newton. In Ocata the default network
backend in jobs is going to be neutron and we need to start
removing jobs that use nova-network from Ocata runs. Since
gate-tempest-dsvm-full-ubuntu-xenial is in the integrated gate
it's a global job and this is a global cap so that it only
runs on stable/newton changes. As noted in the change, we'll
still have nova-network coverage in the cells v1 job in Ocata.

Change-Id: I3c13cb525da126e272789d0db6301f5e58c13b51