Updated openstack/openstack Project: openstack/python-cinderclient  a01a3e1b7f6dd0919ffa15a4d54b84f0ac1e7915
Move old oslo-incubator code out of openstack/common

As part of the first community-wide goal, teams were asked to
remove the openstack/common package of their projects if one
existed. This was a byproduct of the old oslo-incubator form
of syncing common functionality.

The package, apiclient, was moved to a top level location and
cliutils was moved to the common module. There are no oslo
specific libraries, the recommended solution is to move it in
tree and maintain it there.

Change-Id: Iee52004bd33c19d63133577ff466164b85fd6ca6