Updated openstack/openstack Project: openstack/nova  044e9f7ab0cd1b760378cdd02307094dcd05b950
pci: remove pci device from claims and allocations when freeing it

In drop_move_claim we call free pci device when we need to drop a
specific device from the src/dest node. This is done by calling
pci manager free_device. The current code just update the device
status to available in database but doesn't remove it from the
pci manager claims and allocations lists. This patch adds the
removal as well.

Closes-Bug: #1622854

Change-Id: If1cd6f3a635759cd55d116a34ca164630c61e085