Updated openstack/openstack Project: openstack/tempest  853b84a76dd48f17296c951fabed4dba125d2ad8
Merge the separated link lines on compute client

On current docstring of service clients, every link to API
reference is separated to 2 lines. That is unreadable on the
code and the documentation which is created from the docstring
also will be unreadablebecause of the separated link lines.

This patch merges the separated lines into a single line for the
readability.

Change-Id: I14bc0f75bb8f149ddee1d20f6ac7129756244f93