Updated openstack/openstack Project: openstack-infra/shade  b36f0acd582c85e873e60f1f90847a498104d646
Add unit tests for image and flavor normalization

Also, while we're at it - it turns out we were NOT doing normalization
in any meaningful way on nova-based images. This is ok - I am completely
unaware of any clouds that do not expose the glance API any more. But we
should fix it.

Change-Id: I4a87bb9b83fcf12ad25ec211c5d13fcf34432793