Updated openstack/openstack Project: openstack/nova  ef55e4fd9792b343359b325ee8a27e35d04012d7
Fix database poison warnings, part 23

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings (12 new ones) from:

    nova.tests.unit.virt.ironic.test_driver.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I2f11c85a3fd192d93f010b781af60b78ff6914b9
Related-Bug: #1568414