Updated openstack/openstack Project: openstack/neutron-lbaas  41b46bdf8c2c8e24cc8851a505a027dff4c5ebf2
Remove last vestiges of oslo-incubator

This is a community goal for Ocata.  We're already not using any of
the code, but there are some other files and references left over.

These files used to be installed from oslo-incubator:
  tools/install_venv.py
  tools/install_venv_common.py
  tools/with_venv.sh
They were meant to be used by the legacy run_tests.sh, which does not
exist in this repo.

Change-Id: I45c46dcd53ff2f6eaee4faa6da119d5144166533
Partial-Bug: #1639103