Updated openstack/openstack Project: openstack/neutron  ef4de8a23ce9019f25cb3ee3ae777cdb4597e8b9
Check if namespace exists before getting devices

If a neutron router has no ports defined, its namespace is deleted
by the cronjob /etc/cron.d/neutron-l3-agent-netns-cleanup.
It is required to check if the namespace exists before calling
iproute2 commands against the namespace.
It is not enough for the string of the namespace uuid to be defined,
the namespace must really exist on the network node.
This patch checks if the namespace exists when calling get_devices().
Otherwise the agent log file will be flooded with messages like
Cannot open network namespace "qrouter-<uuid>": No such file or directory
Failed to process compatible router '<uuid>'

Related-bug: 1573073

Change-Id: I744ef11529f9da5cbfdb812de0b35b95f9d078bb