Updated openstack/openstack Project: openstack-infra/project-config  7613e9de5098614a4d6d93ca47b2d5fd04a82191
fix trove ci skip-if

It appears that there can only be one skip-if and you have to figure
out how to encode all of the conditions you want into one skip-if.

Luckily, this is sufficient for trove's usage; skip the job on two
projects if the branches are not stable/liberty/mitaka or newton.

For example, see http://logs.openstack.org/93/382493/4/check/. This
was a change on master and both the legacy and non-legacy jobs got
run. They would pass now because the contents of the two repos have
not diverged; that's not the point, they shouldn't both be run :)

Change-Id: I8609d16b3442374ec1c44ba5b1ec247342b3fe7c